Merged PR 5781438: [202012] Move ACMS to Buster base, and make a Python 2 layer for ACMS and pmon
ACMS container in the 202012 branch is currently based on Ubuntu 18.04. Move it to a Buster base, so that it's the same base container image as the other containers.

Additionally, since both ACMS and pmon are using Python 2, and they're each installing Python 2 in their own containers, make a common Python 2 layer, and have both pmon and ACMS be based off of that container. This saves a good amount of disk space.

Related work items: #13319130