Merged PR 6356040: No longer hardcode production syslog server IP
Cherry-pick PR https://msazure.visualstudio.com/One/_git/Networking-acs-buildimage/pullrequest/4251943 to internal-202012 branch has conflict. So, created a dedicated PR to make that change into internal-202012 branch.

The original description in PR 4251943:
"""
Production syslog server (10.20.6.16) should be present in all minigraphs, so there should no longer be a need to hardcode it.

Hardcoding it here also can cause issues, as lab devices have the potential to spam the production syslog server and overload it.

Also add square brackets around IP address to support IPv6 addresses. This change was added in public a while ago, but it appears the internal patch was never updated to reflect it.
"""

Signed-off-by: Xin Wang <xiwang5@microsoft.com>